chore: Update env.example with new secrets
This commit updates the env.example file to include new secrets: SECRET_KEY and CRON_SECRET. These secrets are necessary for the application to function properly. By adding them to the env.example file, we ensure that developers have a clear template to follow when setting up their local environment.