chore: Update npm dependencies
This commit updates the npm dependencies in the package.json file. It adds the following packages: @upstash/redis, axios, cheerio, get-urls, socks-proxy-agent, typescript-eslint, and zod. It also updates the versions of existing packages. This change ensures that the project is using the latest stable versions of the dependencies, improving performance, security, and compatibility.